+++ public/index.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <title>Template Site</title>
-    </head>
-    <body>
-        <section id='root'></section>
-    </body>
+  <head>
+    <title>Stopwatch</title>
+  </head>
+  <body>
+    <section id="root"></section>
+  </body>
 </html>
