+++ src/App.css
@@ -1,4 +1,68 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap");
+
+body {
+  margin: 0;
+  padding: 0;
+  height: 100vh;
+  font-family: "Roboto", sans-serif;
+}
+
+.container {
+  width: 80%;
+  max-width: 1200px;
+  height: 100vh;
+  margin: 0 auto;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  overflow-y: scroll;
+}
+
+h1 {
+  font-size: 3rem;
+}
+
 .btn-container {
   display: flex;
   gap: 1rem;
+}
+
+button {
+  display: inline-block;
+  padding: 10px 20px;
+  font-size: 16px;
+  text-align: center;
+  text-decoration: none;
+  cursor: pointer;
+  color: #fff;
+  border: none;
+  border-radius: 5px;
+  transition: background-color 0.2s;
+}
+
+button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.start {
+  background-color: rgb(30, 30, 30);
+  color: white;
+}
+
+button:hover:enabled {
+  background-color: #535353;
+}
+
+.stop {
+  background-color: rgb(170, 37, 37);
+}
+
+.reset {
+  background-color: rgb(103, 103, 103);
+}
+
+.lap {
+  background-color: orange;
 }
